add: domains actions read from spicedb schema
